data exploration
the kaggle site has a notebook feature so you can run scipts within the site and get visualizations there without needing to set up a local environment. very nice for exploration. This is just a test I did with a given tutorial from the competition. I will go back and modify and change it up a bit.